Changed currAddress from int to int16_t as all memory addresses can be fit into 16 bits.

+++ src/assembler/encodeInstructions.c
@@ -2,8 +2,8 @@
 
 static int32_t encodeDataProcessing(instr_t *instr);
 static int32_t encodeMultiply(instr_t *instr);
-static int32_t encodeSingleDataTransfer(instr_t *instr, int currAddress);
-static int32_t encodeBranch(instr_t *instr, int currAddress);
+static int32_t encodeSingleDataTransfer(instr_t *instr, int16_t currAddress);
+static int32_t encodeBranch(instr_t *instr, int16_t currAddress);
 
 // Declare pointer to array of instructions
 // and address counter and symbol table
@@ -60,7 +60,7 @@ int32_t encodeMultiply(instr_t *instr) {
 }
 
 
-int32_t encodeSingleDataTransfer(instr_t *instr, int currAddress) {
+int32_t encodeSingleDataTransfer(instr_t *instr, int16_t currAddress) {
 
     // If SDT expr, then calculate offset, store it in memory
 
@@ -106,7 +106,7 @@ int32_t encodeSingleDataTransfer(instr_t *instr, int currAddress) {
 }
 
 
-int32_t encodeBranch(instr_t *instr, int currAddress) {
+int32_t encodeBranch(instr_t *instr, int16_t currAddress) {
     int16_t addressDiff
             = (int16_t) ((instr->targetAddress * WORD_LENGTH) - currAddress);
     // Take into account off-by-8 effect of the ARM pipeline.
@@ -124,7 +124,7 @@ int32_t encodeBranch(instr_t *instr, int currAddress) {
 }
 
 int32_t encode(assIns_t *instr) {
-    static int currAddress = 0;
+    static int16_t currAddress = 0;
     int32_t binaryInstruction = 0;
     instr_t *format = getFormat(instr);
